[fix] GroupRanHistory Entity 수정 및 API 변경 반영

+++ app/src/main/java/com/ssu/better/presentation/component/CircleRankProfile.kt
@@ -18,7 +18,7 @@ import com.ssu.better.ui.theme.BetterColors
 @Composable
 fun CircleRankProfile(
     modifier: Modifier = Modifier,
-    score: Int,
+    score: Long,
 ) {
     Surface(
         modifier = modifier,
